requirements.txt
Django~=2.2.0
django-colorfield==0.1.15
django-tinymce==2.8.0
django-mailer==1.2.6
requests~=2.20.0
mysqlclient==1.3.13
django-widget-tweaks==1.4.3
beautifulsoup4~=4.6.3
django-formtools==2.1
bcrypt
lxml
